move monitoring test to a separate suite
