+++ test/e2e-tests.sh
@@ -30,6 +30,7 @@ ensure_serverless_installed
 # Run Knative Serving & Eventing downstream E2E tests.
 downstream_serving_e2e_tests
 downstream_eventing_e2e_tests
+downstream_monitoring_e2e_tests
 if [[ $TEST_KNATIVE_KAFKA == true ]]; then
   ensure_kafka_no_auth
   downstream_knative_kafka_e2e_tests
