+++ test/e2ekafka/knative_kafka_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/openshift-knative/serverless-operator/test"
 	"github.com/openshift-knative/serverless-operator/test/e2e"
+	"github.com/openshift-knative/serverless-operator/test/monitoringe2e"
 	v1a1test "github.com/openshift-knative/serverless-operator/test/v1alpha1"
 )
 
@@ -48,11 +49,11 @@ func TestKnativeKafka(t *testing.T) {
 
 	t.Run("verify health metrics work correctly", func(t *testing.T) {
 		// Eventing should be up
-		if err := e2e.VerifyHealthStatusMetric(caCtx, "eventing_status", "1"); err != nil {
+		if err := monitoringe2e.VerifyHealthStatusMetric(caCtx, "eventing_status", "1"); err != nil {
 			t.Fatal("Failed to verify that health metrics work correctly for Eventing", err)
 		}
 		// KnativeKafka should be up
-		if err := e2e.VerifyHealthStatusMetric(caCtx, "kafka_status", "1"); err != nil {
+		if err := monitoringe2e.VerifyHealthStatusMetric(caCtx, "kafka_status", "1"); err != nil {
 			t.Fatal("Failed to verify that health metrics work correctly for KnativeKafka", err)
 		}
 	})
